+++ src/api/brekekejs.d.ts
@@ -212,6 +212,7 @@ export type PbxGetProductInfoRes = {
   'webphone.turn.server': string
   'webphone.turn.username': string
   'webphone.uc.host': string
+  'webphone.useragent': string
   'webphone.users.max': string
   'webrtcclient.dtmfSendMode': string
 }